+++ src/ForWebforms.Tests/MvcBridgeTests.cs
@@ -7,6 +7,7 @@ using System.Web.Routing;
 using System.Web.UI;
 using Moq;
 using Shouldly;
+using ExitStrategy.ForWebforms.Bridge;
 
 namespace ExitStrategy.ForWebforms.Tests
 {
