+++ src/ForWebforms/ForWebforms.csproj
@@ -67,7 +67,7 @@
   <ItemGroup>
     <Compile Include="Display.cs" />
     <Compile Include="Editor.cs" />
-    <Compile Include="HttpRequestProvider.cs" />
+    <Compile Include="Bridge\HttpRequestProvider.cs" />
     <Compile Include="ModelBinding\DataItemContainerBindingStrategy.cs" />
     <Compile Include="ModelBinding\DataSourceBindingStrategy.cs" />
     <Compile Include="ModelBinding\EnumerableBasedBindingStrategy.cs" />
@@ -79,14 +79,14 @@
     <Compile Include="ModelBinding\BindingStrategySelector.cs" />
     <Compile Include="ModelBinding\ModelDefinition.cs" />
     <Compile Include="ModelBinding\ModelValueExtractor.cs" />
-    <Compile Include="MvcBridge.cs" />
+    <Compile Include="Bridge\MvcBridge.cs" />
     <Compile Include="MvcControl.cs" />
     <Compile Include="Version.cs" />
-    <Compile Include="WebformsController.cs" />
+    <Compile Include="Bridge\WebformsController.cs" />
     <Compile Include="Partial.cs" />
     <Compile Include="NewObjectExpressionBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="WebformsViewDataContainer.cs" />
+    <Compile Include="Bridge\WebformsViewDataContainer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="ForWebforms.nuspec">
