+++ src/ForWebforms/MvcControl.cs
@@ -1,3 +1,4 @@
+using ExitStrategy.ForWebforms.Bridge;
 using ExitStrategy.ForWebforms.ModelBinding;
 using System;
 using System.Collections;
