+++ src/ForWebforms/Version.cs
@@ -4,6 +4,7 @@
     {
         public const string VersionNumber = "0.0.2";
         private const bool IsPreRelease = true;
+
         public const string SemverVersionNumber = VersionNumber + (IsPreRelease ? "-alfa" : "");
     }
 }
